refactor: 이미지정보 imageSize로 class 분리 작업
